backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq  # Add any other dependencies here
python-multipart
